Average number of messages per client

+++ SmsFeedback_Take4/Models/Helpers/RepInfoBox.cs
@@ -7,10 +7,10 @@ namespace SmsFeedback_Take4.Models.Helpers
 {
     public class RepInfoBox
     {
-        public int value;
+        public double value;
         public string unit;
 
-        public RepInfoBox(int iValue, string iUnit)
+        public RepInfoBox(double iValue, string iUnit)
         {
             value = iValue;
             unit = iUnit;
